refactor(user): 닉네임 관련 로직 UserRepository로 통합
- 불필요한 `NicknameRepository`와 `NicknameRepositoryImpl`을 삭제합니다.
- 닉네임 조회, 저장, 관찰 관련 함수들을 `UserRepository` 및 `UserRepositoryImpl`으로 이동시킵니다.
- 닉네임 관련 유스케이스들(`SaveNicknameUseCase`, `GetNicknameUseCase`, `ObserveNicknameUseCase`)의 의존성을 `NicknameRepository`에서 `UserRepository`로 변경합니다.
- `RepositoryModule`에서 `NicknameRepository`의 바인딩 코드를 제거합니다.
- 유저 프로필 정보를 동기화하는 `SyncUserUseCase`를 추가합니다.
