+++ data/src/main/java/com/smtm/pickle/data/di/RepositoryModule.kt
@@ -3,13 +3,11 @@ package com.smtm.pickle.data.di
 import com.smtm.pickle.data.repository.AuthRepositoryImpl
 import com.smtm.pickle.data.repository.LedgerRepositoryImpl
 import com.smtm.pickle.data.repository.MateRepositoryImpl
-import com.smtm.pickle.data.repository.NicknameRepositoryImpl
 import com.smtm.pickle.data.repository.VerdictRepositoryImpl
 import com.smtm.pickle.data.repository.UserRepositoryImpl
 import com.smtm.pickle.domain.repository.AuthRepository
 import com.smtm.pickle.domain.repository.LedgerRepository
 import com.smtm.pickle.domain.repository.MateRepository
-import com.smtm.pickle.domain.repository.NicknameRepository
 import com.smtm.pickle.domain.repository.VerdictRepository
 import com.smtm.pickle.domain.repository.UserRepository
 import dagger.Binds
@@ -28,10 +26,6 @@ abstract class RepositoryModule {
     @Binds
     @Singleton
     abstract fun bindUserRepository(impl: UserRepositoryImpl): UserRepository
-
-    @Binds
-    @Singleton
-    abstract fun bindNicknameRepository(impl: NicknameRepositoryImpl): NicknameRepository
 
     @Binds
     @Singleton
