--- domain/src/main/java/com/smtm/pickle/domain/repository/NicknameRepository.kt
@@ -1,12 +1,0 @@
-package com.smtm.pickle.domain.repository
-
-import kotlinx.coroutines.flow.Flow
-
-interface NicknameRepository {
-
-    suspend fun saveNickname(nickname: String)
-
-    suspend fun getNickname(): String
-
-    fun observeNickname(): Flow<String?>
-}
