+++ domain/src/main/java/com/smtm/pickle/domain/usecase/nickname/GetNicknameUseCase.kt
@@ -1,13 +1,13 @@
 package com.smtm.pickle.domain.usecase.nickname
 
 import com.smtm.pickle.domain.common.utils.runSuspendCatching
-import com.smtm.pickle.domain.repository.NicknameRepository
+import com.smtm.pickle.domain.repository.UserRepository
 import javax.inject.Inject
 
 class GetNicknameUseCase @Inject constructor(
-    private val nicknameRepository: NicknameRepository
+    private val userRepository: UserRepository
 ) {
     suspend operator fun invoke(): Result<String> = runSuspendCatching {
-        nicknameRepository.getNickname()
+        userRepository.getNickname()
     }
 }
