+++ domain/src/main/java/com/smtm/pickle/domain/usecase/nickname/ObserveNicknameUseCase.kt
@@ -1,11 +1,11 @@
 package com.smtm.pickle.domain.usecase.nickname
 
-import com.smtm.pickle.domain.repository.NicknameRepository
+import com.smtm.pickle.domain.repository.UserRepository
 import kotlinx.coroutines.flow.Flow
 import javax.inject.Inject
 
 class ObserveNicknameUseCase @Inject constructor(
-    private val nicknameRepository: NicknameRepository
+    private val userRepository: UserRepository
 ) {
-    operator fun invoke(): Flow<String?> = nicknameRepository.observeNickname()
+    operator fun invoke(): Flow<String?> = userRepository.observeNickname()
 }
