+++ domain/src/main/java/com/smtm/pickle/domain/usecase/nickname/SyncUserUseCase.kt
@@ -4,10 +4,10 @@ import com.smtm.pickle.domain.common.utils.runSuspendCatching
 import com.smtm.pickle.domain.repository.UserRepository
 import javax.inject.Inject
 
-class SaveNicknameUseCase @Inject constructor(
+class SyncUserUseCase @Inject constructor(
     private val userRepository: UserRepository
 ) {
-    suspend operator fun invoke(nickname: String) = runSuspendCatching {
-        userRepository.saveNickname(nickname)
+    suspend operator fun invoke(): Result<String> = runSuspendCatching {
+        userRepository.getNickname()
     }
 }
